@@ -15,27 +15,29 @@ gem "font-awesome-rails"
 gem "haml-rails"
 gem "high_voltage"
 gem "inifile"
-gem "jquery-rails", "~> 4.2.0"
+gem "jquery-rails", ">= 4.2.0"
 gem "neat"
 gem "normalize-rails"
 gem "octokit"
 gem "omniauth-github"
-gem "paranoia", "~> 2.0"
+gem "paranoia", "~> 2.2"
 gem "pg"
 gem "puma"
-gem "rails", "4.2.7.1"
+gem "rails", "5.0.0.1"
 gem "rails-assets-classnames", source: "https://rails-assets.org"
 gem "rails-assets-es5-shim", source: "https://rails-assets.org"
 gem "rails-assets-lodash", source: "https://rails-assets.org"
 gem "react-rails"
-gem "resque", "~> 1.25.0"
+gem "resque", ">= 1.25.0"
 gem "resque-scheduler"
 gem "resque-sentry"
 gem "rest-client", ">= 1.8.0"
 gem "sass-rails"
+gem "sinatra", "2.0.0.beta2"
 gem "split", require: "split/dashboard"
 gem "stripe"
 gem "uglifier", ">= 2.7.2"
+gem "record_tag_helper"
 
 group :staging, :production do
   gem "rack-timeout"
@@ -48,12 +50,11 @@ group :development, :test do
   gem "byebug"
   gem "dotenv-rails"
   gem "foreman"
-  gem "quiet_assets"
   gem "rspec-rails", ">= 3.4"
 end
 
 group :test do
-  gem "capybara", "~> 2.4.0"
+  gem "capybara", ">= 2.4.0"
   gem "capybara-webkit", "~> 1.6"
   gem "database_cleaner"
   gem "factory_girl_rails"
