@@ -1,6 +1,7 @@
-require File.expand_path('../boot', __FILE__)
+require File.expand_path("../boot", __FILE__)
+require File.expand_path("lib/redirect_to_configuration")
 
-require 'rails/all'
+require "rails/all"
 
 Bundler.require(:default, Rails.env)
 
@@ -11,10 +12,10 @@ module Houndapp
     config.filter_parameters += [:password]
     config.active_support.escape_html_entities_in_json = true
     config.active_job.queue_adapter = :resque
-    config.middleware.insert_before "Rack::ETag", "Rack::Deflater"
+    config.middleware.insert_before Rack::ETag, Rack::Deflater
     config.middleware.insert_before(
-      "Rack::ETag",
-      "RedirectToConfiguration"
+      Rack::ETag,
+      RedirectToConfiguration,
     )
     config.exceptions_app = routes
 
