@@ -9,8 +9,10 @@ Houndapp::Application.configure do
   config.cache_classes = true
 
   # Configure static asset server for tests with Cache-Control for performance
-  config.serve_static_files = true
-  config.static_cache_control = "public, max-age=3600"
+  config.public_file_server.enabled = true
+  config.public_file_server.headers = {
+    "Cache-Control" => "public, max-age=3600",
+  }
 
   # Show full error reports and disable caching
   config.consider_all_requests_local       = true
