@@ -14,7 +14,7 @@ describe ActivationsController, "#create" do
         scope: membership.user,
       ).to_json
 
-      post :create, repo_id: repo.id, format: :json
+      post :create, params: { repo_id: repo.id }, format: :json
 
       expect(response).to have_http_status(:created)
       expect(response.body).to eq expected_repo_json
@@ -38,7 +38,7 @@ describe ActivationsController, "#create" do
         allow(RepoActivator).to receive(:new).and_return(activator)
         stub_sign_in(membership.user)
 
-        post :create, repo_id: repo.id, format: :json
+        post :create, params: { repo_id: repo.id }, format: :json
 
         response_body = JSON.parse(response.body)
         expect(response.code).to eq "502"
@@ -60,7 +60,7 @@ describe ActivationsController, "#create" do
       allow(RepoActivator).to receive(:new).and_return(activator)
       stub_sign_in(membership.user)
 
-      post :create, repo_id: repo.id, format: :json
+      post :create, params: { repo_id: repo.id }, format: :json
 
       expect(analytics).to have_tracked("Repo Activation Failed").
         for_user(membership.user).
@@ -82,9 +82,8 @@ describe ActivationsController, "#create" do
       allow(RepoActivator).to receive(:new).and_return(activator)
       stub_sign_in(user)
 
-      expect { post :create, repo_id: repo.id, format: :json }.to raise_error(
-        ActivationsController::CannotActivatePaidRepo
-      )
+      expect { post :create, params: { repo_id: repo.id }, format: :json }.
+        to raise_error(ActivationsController::CannotActivatePaidRepo)
       expect(activator).not_to have_received(:activate)
     end
   end
