@@ -27,7 +27,7 @@ end
 
 Capybara.configure do |config|
   config.javascript_driver = :webkit
-  config.default_wait_time = 4
+  config.default_max_wait_time = 4
 end
 
 Capybara::Webkit.configure(&:block_unknown_urls)
